refactor: 기존 스낵게임 legacy 폴더 이동
