+++ src/pages/games/SnackGame/game/SnackGameBase.tsx
