+++ src/pages/games/SnackGame/game/hook/usePixiCanvas.tsx
