+++ src/pages/games/SnackGame/game/legacy/components/ExpBarChart.tsx
