+++ src/pages/games/SnackGame/game/legacy/components/GameResult.tsx
