+++ src/pages/games/SnackGame/game/legacy/components/SnackGameHUD.tsx
