+++ src/pages/games/SnackGame/game/legacy/model/appleGame/apple.ts
