+++ src/pages/games/SnackGame/game/legacy/model/appleGame/goldenApple.ts
