+++ src/pages/games/SnackGame/game/legacy/model/appleGame/plainApple.ts
