+++ src/pages/games/SnackGame/game/legacy/model/common/click.ts
