+++ src/pages/games/SnackGame/game/legacy/model/common/drag.ts
