+++ src/pages/games/SnackGame/game/legacy/model/common/particle.ts
