+++ src/pages/games/SnackGame/game/legacy/model/snackGame/goldSnack.ts
