+++ src/pages/games/SnackGame/game/legacy/model/snackGame/plainSnack.ts
