+++ src/pages/games/SnackGame/game/legacy/model/snackGame/snack.ts
