+++ src/pages/games/SnackGame/game/legacy/model/snackGame/snackGame.ts
