+++ src/pages/games/SnackGame/game/legacy/model/snackGame/snackGameInf.ts
