+++ src/pages/games/SnackGame/game/legacy/model/snackGame/snackGameMobile.ts
