+++ src/pages/games/SnackGame/game/legacy/view/appleGame/DefaultMode.tsx
@@ -2,16 +2,16 @@ import { useEffect, useState } from 'react';
 
 import QueryBoundary from '@components/base/QueryBoundary';
 import retryError from '@components/Error/RetryError';
-import GameResult from '@pages/games/SnackGame/game/components/GameResult';
+import GameResult from '@pages/games/SnackGame/game/legacy/components/GameResult';
 import {
   goldModAppleType,
   scoredAppleRectType,
 } from '@pages/games/SnackGame/game/game.type';
-import Apple from '@pages/games/SnackGame/game/model/appleGame/apple';
-import { AppleGame } from '@pages/games/SnackGame/game/model/appleGame/appleGame';
-import { GoldenApple } from '@pages/games/SnackGame/game/model/appleGame/goldenApple';
-import PlainApple from '@pages/games/SnackGame/game/model/appleGame/plainApple';
-import SnackGameHUD from '@pages/games/SnackGame/game/components/SnackGameHUD';
+import Apple from '@pages/games/SnackGame/game/legacy/model/appleGame/apple';
+import { AppleGame } from '@pages/games/SnackGame/game/legacy/model/appleGame/appleGame';
+import { GoldenApple } from '@pages/games/SnackGame/game/legacy/model/appleGame/goldenApple';
+import PlainApple from '@pages/games/SnackGame/game/legacy/model/appleGame/plainApple';
+import SnackGameHUD from '@pages/games/SnackGame/game/legacy/components/SnackGameHUD';
 
 import {
   useAppleGameRefresh,
