+++ src/pages/games/SnackGame/game/screen/GameScreen.ts
