+++ src/pages/games/SnackGame/game/screen/LoadScreen.ts
