+++ src/pages/games/SnackGame/game/ui/Cloud.ts
