+++ src/pages/games/SnackGame/game/ui/Label.ts
