+++ src/pages/games/SnackGame/game/ui/SnackGameLetter.ts
