+++ src/pages/games/SnackGame/game/ui/SnackGameLogo.ts
