+++ src/pages/games/SnackGame/game/ui/Waves.ts
