+++ src/pages/games/SnackGame/game/util/assets.ts
