+++ src/pages/games/SnackGame/game/util/getUrlParams.ts
