+++ src/pages/games/SnackGame/game/util/navigation.ts
