+++ src/pages/games/SnackGame/game/util/random.ts
